requirements.txt
pandas
numpy
matplotlib
scikit-learn
ipykernel
keras
seaborn
kaggle
pillow
opencv-python
scikit-image
